cleanup + move code into modules
