+++ relay-server/src/processing/replays/filter.rs
@@ -1,8 +1,10 @@
 use relay_dynamic_config::Feature;
+use relay_statsd::metric;
 
 use crate::managed::{Managed, Rejected};
 use crate::processing::Context;
-use crate::processing::replays::{Error, SerializedReplays};
+use crate::processing::replays::{Error, ExpandedReplays, SerializedReplays};
+use crate::statsd::RelayCounters;
 
 /// Reject data if the feature is disabled.
 pub fn feature_flag(
@@ -18,4 +20,44 @@ pub fn feature_flag(
         true => Err(replays.reject_err(Error::FilterFeatureFlag)),
         false => Ok(replays),
     }
+}
+
+pub fn filter(replays: &mut Managed<ExpandedReplays>, ctx: Context<'_>) {
+    let client_addr = replays.headers.meta().client_addr();
+    let event_id = replays.headers.event_id();
+
+    replays.retain(
+        |replays| &mut replays.replays,
+        |replay, _| {
+            let Some(event) = replay.get_event().value() else {
+                return Err(Error::NoEventContent);
+            };
+
+            relay_filter::should_filter(
+                event,
+                client_addr,
+                &ctx.project_info.config.filter_settings,
+                ctx.global_config.filters(),
+            )
+            .map_err(Error::Filtered)?;
+
+            // Log segments that exceed the hour limit so we can diagnose errant SDKs
+            // or exotic customer implementations.
+            if let Some(segment_id) = event.segment_id.value()
+                && *segment_id > 720
+            {
+                metric!(counter(RelayCounters::ReplayExceededSegmentLimit) += 1);
+
+                relay_log::debug!(
+                    event_id = ?event_id,
+                    project_id = ctx.project_info.project_id.map(|v| v.value()),
+                    organization_id = ctx.project_info.organization_id.map(|o| o.value()),
+                    segment_id = segment_id,
+                    "replay segment-exceeded-limit"
+                );
+            }
+
+            Ok::<_, Error>(())
+        },
+    )
 }
