+++ relay-server/src/processing/replays/mod.rs
@@ -1,55 +1,63 @@
 use std::sync::Arc;
 
-use relay_dynamic_config::Feature;
-use relay_event_normalization::{GeoIpLookup, RawUserAgentInfo, replay};
-use relay_event_schema::processor::{self, ProcessingState};
+use bytes::Bytes;
+
+use relay_event_normalization::GeoIpLookup;
 use relay_event_schema::protocol::Replay;
 use relay_filter::FilterStatKey;
-use relay_pii::{PiiConfigError, PiiProcessor};
+use relay_pii::PiiConfigError;
 use relay_protocol::Annotated;
 use relay_quotas::{DataCategory, RateLimits};
-use relay_replays::recording::RecordingScrubber;
-use relay_statsd::metric;
 
-use crate::envelope::{ContentType, EnvelopeHeaders, Item, ItemHeaders, ItemType, Items};
+use crate::envelope::{EnvelopeHeaders, Item, ItemHeaders, ItemType};
 use crate::managed::{Counted, Managed, ManagedEnvelope, OutcomeError, Rejected};
-
-use crate::Envelope;
-use crate::processing::{self, Context, CountRateLimited, Forward, Output, QuotaRateLimiter};
+use crate::processing::{self, Context, CountRateLimited, Output, QuotaRateLimiter};
 use crate::services::outcome::{DiscardReason, Outcome};
-use crate::services::processor::replay::ReplayVideoEvent;
-use crate::statsd::{RelayCounters, RelayTimers};
-
-use bytes::Bytes;
 
 pub type Result<T, E = Error> = std::result::Result<T, E>;
 
 mod filter;
+mod forward;
 mod process;
+mod validate;
 
-// FIXME: Come up with more specific error variants and how to map them.
-// FIXME: Check if we already have errors that clone
-#[derive(Debug, thiserror::Error, Clone)]
+#[derive(Debug, thiserror::Error)]
 pub enum Error {
-    #[error("rate limited")]
-    RateLimited(RateLimits),
-    // Q: When to do this vs: FeatureDisabled(Feature)
+    /// Replays filtered because of a missing feature flag.
     #[error("replay feature flag missing")]
     FilterFeatureFlag,
 
+    // TODO:: Q: Add a more detailed version here, for the (0,n) case?
+    /// There is an invalid amount of `replay_event` and `replay_recording` items in the envelope.
+    ///
+    /// There should be either 0 of both or 1 of both.
+    #[error("invalid item count")]
+    InvalidItemCount,
+
     /// The Replay event could not be parsed from JSON.
     #[error("invalid json: {0}")]
-    CouldNotParse(String),
-
-    #[error("invalid replay video")]
-    InvalidReplayVideoEvent,
-
-    #[error("invalid replay")]
-    InvalidReplayRecordingEvent,
+    CouldNotParseEvent(String),
 
     /// The Replay event was parsed but did not match the schema.
     #[error("no data found")]
     NoEventContent,
+
+    /// The Replay contains invalid data or is missing a required field.
+    #[error("invalid payload {0}")]
+    InvalidPayload(String),
+
+    /// The Replay video could not be parsed.
+    #[error("invalid replay video")]
+    InvalidReplayVideoEvent,
+
+    /// The processing of the Replay Recording failed.
+    #[error("invalid replay")]
+    InvalidReplayRecordingEvent,
+
+    // FIXME: Think about merging these since having both seems a bit overkill.
+    /// The PII config for scrubbing the recording could not be loaded.
+    #[error("invalid pii config")]
+    PiiConfigError(PiiConfigError),
 
     /// An error occurred during PII scrubbing of the Replay.
     ///
@@ -57,15 +65,15 @@ pub enum Error {
     #[error("failed to scrub PII: {0}")]
     CouldNotScrub(String),
 
-    #[error("invalid item count")]
-    InvalidItemCount,
+    /// The Replays are rate limited.
+    #[error("rate limited")]
+    RateLimited(RateLimits),
 
+    /// Replays filtered due to a filtering rule
     #[error("replay filtered with reason: {0:?}")]
-    ReplayFiltered(FilterStatKey),
+    Filtered(FilterStatKey),
 
-    #[error("invalid pii config")]
-    PiiConfigError(PiiConfigError),
-
+    /// Failed to re-serialize the replay event.
     #[error("failed to serialize replay")]
     FailedToSerializeReplayEvent,
 }
@@ -74,33 +82,31 @@ impl OutcomeError for Error {
     type Error = Self;
 
     fn consume(self) -> (Option<Outcome>, Self::Error) {
-        // Q: Is it ok to add some logging here?
         let outcome = match &self {
-            Self::RateLimited(limits) => {
-                let reason_code = limits.longest().and_then(|limit| limit.reason_code.clone());
-                Some(Outcome::RateLimited(reason_code))
-            }
-            // Q: When to do this vs: Outcome::Invalid(DiscardReason::FeatureDisabled(_))
-            Self::FailedToSerializeReplayEvent => None,
             Self::FilterFeatureFlag => None,
-            Self::CouldNotParse(_) => Some(Outcome::Invalid(DiscardReason::InvalidReplayEvent)),
-            Self::InvalidReplayVideoEvent => {
-                Some(Outcome::Invalid(DiscardReason::InvalidReplayVideoEvent))
+            Self::InvalidItemCount => Some(Outcome::Invalid(DiscardReason::DuplicateItem)),
+            Self::CouldNotParseEvent(_) => {
+                Some(Outcome::Invalid(DiscardReason::InvalidReplayEvent))
             }
             Self::NoEventContent => {
                 Some(Outcome::Invalid(DiscardReason::InvalidReplayEventNoPayload))
             }
-            Self::CouldNotScrub(_) => Some(Outcome::Invalid(DiscardReason::InvalidReplayEventPii)),
-            // TODO: Not super happy with this
-            // (0,1) (1,0) => mismatch (check how attachments does it)
-            // (1,1)
-            // (n,n) => duplicate
-            Self::InvalidItemCount => Some(Outcome::Invalid(DiscardReason::DuplicateItem)),
-            Self::ReplayFiltered(key) => Some(Outcome::Filtered(key.clone())),
-            Self::PiiConfigError(_) => Some(Outcome::Invalid(DiscardReason::ProjectStatePii)),
+            Self::InvalidPayload(_) => Some(Outcome::Invalid(DiscardReason::InvalidReplayEvent)),
+            Self::InvalidReplayVideoEvent => {
+                Some(Outcome::Invalid(DiscardReason::InvalidReplayVideoEvent))
+            }
             Self::InvalidReplayRecordingEvent => {
                 Some(Outcome::Invalid(DiscardReason::InvalidReplayRecordingEvent))
             }
+            Self::PiiConfigError(_) => Some(Outcome::Invalid(DiscardReason::ProjectStatePii)),
+            Self::CouldNotScrub(_) => Some(Outcome::Invalid(DiscardReason::InvalidReplayEventPii)),
+
+            Self::RateLimited(limits) => {
+                let reason_code = limits.longest().and_then(|limit| limit.reason_code.clone());
+                Some(Outcome::RateLimited(reason_code))
+            }
+            Self::Filtered(key) => Some(Outcome::Filtered(key.clone())),
+            Self::FailedToSerializeReplayEvent => Some(Outcome::Invalid(DiscardReason::Internal)),
         };
         (outcome, self)
     }
@@ -174,299 +180,22 @@ impl processing::Processor for ReplaysProcessor {
         replays: Managed<Self::UnitOfWork>,
         ctx: Context<'_>,
     ) -> Result<Output<Self::Output>, Rejected<Self::Error>> {
-        // FIXME: Does it make sense to check for ctx.processing here?
+        // TODO: Q: Does it make sense to check for ctx.processing here?
         let replays = filter::feature_flag(replays, ctx)?;
         let mut replays = process::expand(replays);
 
-        validate(&mut replays);
-        normalize(&mut replays, &self.geoip_lookup);
-        scrub(&mut replays, ctx);
-        filter(&mut replays, ctx);
-        processing(&mut replays, ctx);
+        validate::validate(&mut replays);
+        process::normalize(&mut replays, &self.geoip_lookup);
+        process::scrub(&mut replays, ctx);
+        filter::filter(&mut replays, ctx);
+        process::scrub_recording(&mut replays, ctx);
 
         let replays = self.limiter.enforce_quotas(replays, ctx).await?;
         Ok(Output::just(ReplaysOutput(replays)))
     }
 }
 
-pub fn validate(replays: &mut Managed<ExpandedReplays>) {
-    replays.retain(
-        |replays| &mut replays.replays,
-        |replay, _| {
-            // TODO: Update the error on this to be in line with the actual ones.
-            match replay.get_event().value() {
-                Some(event) => replay::validate(event).map_err(|_| Error::NoEventContent),
-                None => Err(Error::NoEventContent),
-            }
-        },
-    )
-}
-
-pub fn normalize(replays: &mut Managed<ExpandedReplays>, geoip_lookup: &GeoIpLookup) {
-    // Q: We now construct our config here on demand, is that ok, or do we want our own config?
-    let meta = replays.headers.meta();
-    let client_addr = meta.client_addr();
-    let user_agent = RawUserAgentInfo {
-        user_agent: meta.user_agent().map(String::from),
-        client_hints: meta.client_hints().to_owned(),
-    };
-
-    replays.modify(|replay, _| {
-        for replay in replay.replays.iter_mut() {
-            replay::normalize(
-                replay.get_event(),
-                client_addr,
-                &user_agent.as_deref(), // FIXME: Get rid of this.
-                geoip_lookup,
-            )
-        }
-    })
-}
-
-pub fn scrub_event(event: &mut Annotated<Replay>, ctx: Context<'_>) -> Result<(), Error> {
-    if let Some(ref config) = ctx.project_info.config.pii_config {
-        let mut processor = PiiProcessor::new(config.compiled());
-        processor::process_value(event, &mut processor, ProcessingState::root())
-            .map_err(|e| Error::CouldNotScrub(e.to_string()))?;
-    }
-
-    let pii_config = ctx
-        .project_info
-        .config
-        .datascrubbing_settings
-        .pii_config()
-        .map_err(|e| Error::CouldNotScrub(e.to_string()))?;
-
-    if let Some(config) = pii_config {
-        let mut processor = PiiProcessor::new(config.compiled());
-        processor::process_value(event, &mut processor, ProcessingState::root())
-            .map_err(|e| Error::CouldNotScrub(e.to_string()))?;
-    };
-    Ok(())
-}
-
-pub fn scrub(replays: &mut Managed<ExpandedReplays>, ctx: Context<'_>) {
-    replays.retain(
-        |replays| &mut replays.replays,
-        |replay, _| scrub_event(replay.get_event(), ctx),
-    );
-}
-
-pub fn filter(replays: &mut Managed<ExpandedReplays>, ctx: Context<'_>) {
-    let client_addr = replays.headers.meta().client_addr();
-    let event_id = replays.headers.event_id();
-
-    replays.retain(
-        |replays| &mut replays.replays,
-        |replay, _| {
-            // FIXME: technically should never happen since we already check this earlier.
-            let Some(event) = replay.get_event().value() else {
-                return Err(Error::NoEventContent);
-            };
-
-            relay_filter::should_filter(
-                event,
-                client_addr,
-                &ctx.project_info.config.filter_settings,
-                ctx.global_config.filters(),
-            )
-            .map_err(Error::ReplayFiltered)?;
-
-            // Log segments that exceed the hour limit so we can diagnose errant SDKs
-            // or exotic customer implementations.
-            if let Some(segment_id) = event.segment_id.value()
-                && *segment_id > 720
-            {
-                metric!(counter(RelayCounters::ReplayExceededSegmentLimit) += 1);
-
-                relay_log::debug!(
-                    event_id = ?event_id,
-                    project_id = ctx.project_info.project_id.map(|v| v.value()),
-                    organization_id = ctx.project_info.organization_id.map(|o| o.value()),
-                    segment_id = segment_id,
-                    "replay segment-exceeded-limit"
-                );
-            }
-
-            Ok::<_, Error>(())
-        },
-    )
-}
-
-pub fn processing(replays: &mut Managed<ExpandedReplays>, ctx: Context<'_>) {
-    // FIXME: Q: I think ideally we would want to move this check up to the other filter checks no?
-    if !ctx
-        .project_info
-        .has_feature(Feature::SessionReplayRecordingScrubbing)
-    {
-        return;
-    }
-
-    let event_id = replays.headers.event_id();
-
-    replays.retain(
-        |replays| &mut replays.replays,
-        |replay, _| {
-            let datascrubbing_config = ctx
-                .project_info
-                .config
-                .datascrubbing_settings
-                .pii_config()
-                .map_err(|e| Error::PiiConfigError(e.clone()))?
-                .as_ref();
-
-            let mut scrubber = RecordingScrubber::new(
-                ctx.config.max_replay_uncompressed_size(),
-                ctx.project_info.config.pii_config.as_ref(),
-                datascrubbing_config,
-            );
-
-            if scrubber.is_empty() {
-                return Ok::<(), Error>(());
-            }
-
-            let payload = replay.get_recording();
-            *payload = metric!(timer(RelayTimers::ReplayRecordingProcessing), {
-                scrubber.process_recording(payload)
-            })
-            .map(Into::into)
-            .map_err(|error| {
-                relay_log::debug!(
-                    error = &error as &dyn std::error::Error,
-                    event_id = ?event_id,
-                    project_id = ctx.project_info.project_id.map(|v| v.value()),
-                    organization_id = ctx.project_info.organization_id.map(|o| o.value()),
-                    "invalid replay recording"
-                );
-                Error::InvalidReplayRecordingEvent
-            })?;
-            Ok::<(), Error>(())
-        },
-    );
-}
-
-#[derive(Debug)]
-pub struct ReplaysOutput(Managed<ExpandedReplays>);
-
-impl Forward for ReplaysOutput {
-    fn serialize_envelope(
-        self,
-        _ctx: processing::ForwardContext<'_>,
-    ) -> Result<Managed<Box<Envelope>>, Rejected<()>> {
-        Ok(self.0.map(|s, r| {
-            let ExpandedReplays { headers, replays } = s;
-            let mut items = Items::new();
-            let event_id = headers.event_id();
-
-            for replay in replays {
-                match replay {
-                    ExpandedReplay::WebReplay {
-                        event_header,
-                        recording_header,
-                        event,
-                        recording,
-                    } => {
-                        // Parse the replay again.
-                        match event.to_json() {
-                            Ok(json) => {
-                                let event: Bytes = json.into_bytes().into();
-
-                                // items.push(Item::from_parts(event_header, event));
-                                let mut item = Item::new(ItemType::ReplayEvent);
-                                item.set_payload(ContentType::Json, event);
-                                items.push(item);
-
-                                items.push(Item::from_parts(recording_header, recording));
-                            }
-                            Err(error) => {
-                                relay_log::error!(
-                                    error = &error as &dyn std::error::Error,
-                                    event_id = ?event_id,
-                                    "failed to serialize replay"
-                                );
-
-                                // FIXME: Not a super fan of this.
-                                r.reject_err(
-                                    Error::FailedToSerializeReplayEvent,
-                                    ExpandedReplay::WebReplay {
-                                        event_header,
-                                        recording_header,
-                                        event,
-                                        recording,
-                                    },
-                                );
-                            }
-                        }
-                    }
-                    ExpandedReplay::NativeReplay {
-                        video_header,
-                        event,
-                        recording,
-                        video,
-                    } => {
-                        // FIXME: Kind of duplicate logic
-                        let event_bytes: Bytes = match event.to_json() {
-                            Ok(json) => json.into_bytes().into(),
-                            Err(error) => {
-                                relay_log::error!(
-                                    error = &error as &dyn std::error::Error,
-                                    event_id = ?event_id,
-                                    "failed to serialize replay"
-                                );
-
-                                r.reject_err(
-                                    Error::FailedToSerializeReplayEvent,
-                                    ExpandedReplay::NativeReplay {
-                                        video_header,
-                                        event,
-                                        recording,
-                                        video,
-                                    },
-                                );
-                                continue;
-                            }
-                        };
-
-                        let video_event = ReplayVideoEvent {
-                            replay_event: event_bytes,
-                            replay_recording: recording,
-                            replay_video: video,
-                        };
-                        match rmp_serde::to_vec_named(&video_event) {
-                            Ok(payload) => {
-                                items.push(Item::from_parts(video_header, payload.into()));
-                            }
-                            Err(_) => {
-                                // FIXME: Come up with a better error here.
-                                // Err(ProcessingError::InvalidReplay(DiscardReason::InvalidReplayVideoEvent,))
-                                r.reject_err(Error::FailedToSerializeReplayEvent, video_event);
-                            }
-                        }
-                    }
-                }
-            }
-
-            Envelope::from_parts(headers, items)
-        }))
-    }
-
-    #[cfg(feature = "processing")]
-    fn forward_store(
-        self,
-        s: processing::StoreHandle<'_>,
-        ctx: processing::ForwardContext<'_>,
-    ) -> Result<(), Rejected<()>> {
-        // FIXME: Already construct the kafka message here?
-
-        let envelope = self.serialize_envelope(ctx)?;
-        let envelope = ManagedEnvelope::from(envelope).into_processed();
-
-        s.store(crate::services::store::StoreEnvelope { envelope });
-
-        Ok(())
-    }
-}
-
+/// Serialized replays extracted from an envelope.
 #[derive(Debug)]
 pub struct SerializedReplays {
     /// Original envelope headers.
@@ -479,16 +208,6 @@ pub struct SerializedReplays {
     videos: Vec<Item>,
 }
 
-#[derive(Debug)]
-pub struct ExpandedReplays {
-    /// Original envelope headers.
-    headers: EnvelopeHeaders,
-
-    // FIXME: Might not need to be a vec in the future.
-    /// Expanded replays
-    replays: Vec<ExpandedReplay>,
-}
-
 impl Counted for SerializedReplays {
     fn quantities(&self) -> crate::managed::Quantities {
         smallvec::smallvec![(
@@ -498,17 +217,15 @@ impl Counted for SerializedReplays {
     }
 }
 
-impl Counted for ExpandedReplay {
-    fn quantities(&self) -> crate::managed::Quantities {
-        match self {
-            ExpandedReplay::WebReplay { .. } => {
-                smallvec::smallvec![(DataCategory::Replay, 2)]
-            }
-            ExpandedReplay::NativeReplay { .. } => {
-                smallvec::smallvec![(DataCategory::Replay, 1)]
-            }
-        }
-    }
+/// Replays which have been parsed and expanded from their serialized state.
+#[derive(Debug)]
+pub struct ExpandedReplays {
+    /// Original envelope headers.
+    headers: EnvelopeHeaders,
+
+    // FIXME: Might not need to be a vec in the future.
+    /// Expanded replays
+    replays: Vec<ExpandedReplay>,
 }
 
 impl Counted for ExpandedReplays {
@@ -525,7 +242,6 @@ impl Counted for ExpandedReplays {
     }
 }
 
-// FIXME: Better understand why this is needed.
 impl CountRateLimited for Managed<ExpandedReplays> {
     type Error = Error;
 }
@@ -533,6 +249,9 @@ impl CountRateLimited for Managed<ExpandedReplays> {
 #[derive(Debug)]
 // FIXME: Come up with some better naming here.
 // FIXME: Check if we can safely_ignore this warning
+/// An expanded Replay.
+///
+/// Either a web replay, not containing a video or a native replay with a video.
 enum ExpandedReplay {
     WebReplay {
         event_header: ItemHeaders,
@@ -546,6 +265,19 @@ enum ExpandedReplay {
         recording: Bytes,
         video: Bytes,
     },
+}
+
+impl Counted for ExpandedReplay {
+    fn quantities(&self) -> crate::managed::Quantities {
+        match self {
+            ExpandedReplay::WebReplay { .. } => {
+                smallvec::smallvec![(DataCategory::Replay, 2)]
+            }
+            ExpandedReplay::NativeReplay { .. } => {
+                smallvec::smallvec![(DataCategory::Replay, 1)]
+            }
+        }
+    }
 }
 
 impl ExpandedReplay {
@@ -583,3 +315,6 @@ impl ExpandedReplay {
         }
     }
 }
+
+#[derive(Debug)]
+pub struct ReplaysOutput(Managed<ExpandedReplays>);
